test(refactor): move cli specs tp separate folder
